Add files via upload
A Point of Sale (POS) system is the place and process where a sales transaction between a business and a customer is completed. It is the final step in the purchasing journey, where the customer pays for goods or services, and the business records the sale. Traditionally, this was simply a cash register, but modern POS systems are digital, automated, and integrated with other business tools